Clean up making Polynomial a dtype
