+++ bindings/pydrake/BUILD.bazel
@@ -211,6 +211,28 @@ drake_pybind_library(
 )
 
 drake_cc_library(
+    name = "polynomial_types_pybind",
+    hdrs = ["polynomial_types_pybind.h"],
+    declare_installed_headers = 0,
+    visibility = ["//visibility:public"],
+    deps = ["//:drake_shared_library"],
+)
+
+drake_pybind_library(
+    name = "polynomial_py",
+    cc_deps = [
+        ":polynomial_types_pybind",
+        "//bindings/pydrake:documentation_pybind",
+    ],
+    cc_srcs = ["polynomial_py.cc"],
+    package_info = PACKAGE_INFO,
+    py_deps = [
+        ":module_py",
+        "//bindings/pydrake/common:deprecation_py",
+    ],
+)
+
+drake_cc_library(
     name = "symbolic_types_pybind",
     hdrs = ["symbolic_types_pybind.h"],
     declare_installed_headers = 0,
@@ -251,6 +273,7 @@ PY_LIBRARIES_WITH_INSTALL = [
     ":lcm_py",
     ":math_py",
     ":perception_py",
+    ":polynomial_py",
     ":symbolic_py",
     ":trajectories_py",
     "//bindings/pydrake/attic",
@@ -423,6 +446,14 @@ drake_py_unittest(
     name = "perception_test",
     deps = [
         ":perception_py",
+    ],
+)
+
+drake_py_unittest(
+    name = "polynomial_test",
+    deps = [
+        ":polynomial_py",
+        "//bindings/pydrake/common/test_utilities:numpy_compare_py",
     ],
 )
