+++ bindings/pydrake/common/test_utilities/BUILD.bazel
@@ -41,6 +41,7 @@ drake_py_library(
     deps = [
         ":module_py",
         "//bindings/pydrake:autodiffutils_py",
+        "//bindings/pydrake:polynomial_py",
         "//bindings/pydrake:symbolic_py",
     ],
 )
